Bot_trading/requirements.txt
ccxt
pandas
ta
python-dotenv
requests
schedule
yfinance
matplotlib
